backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3.0
python-docx>=1.0.0
reportlab>=4.0.0
pypdf>=3.9.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
dj-database-url>=1.3.0
whitenoise>=6.5.0

